# Optimize distributed.Cache.Get