@tylerwilliams 2025-04-10
adding @jdhollen to help take a look at this code because it's decently critical